Fix output assembly
There was a total failure of the generated assembly due to the binary image being written into a blob. From the Magick++ document: "Caution: if an image format is selected which is capable of supporting fewer colors than the original image or quantization has been requested, the original image will be quantized to fewer colors. Use a copy of the original if this is a problem".